+++ docs/source/changelog.rst
@@ -1,5 +1,9 @@
 Changelog
 =========
+0.7.9
+
+* Channel and listener id generation now relies on uuid4
+
 0.7.7
 
 * Now SocketServer starts correspondant listener on 'l' command